+++ bundle_options_test.go
@@ -8,6 +8,7 @@ import (
 	"github.com/stretchr/testify/require"
 	"golang.org/x/text/language"
 
+	"github.com/vorlif/spreak/catalog"
 	"github.com/vorlif/spreak/internal/util"
 )
 
@@ -112,7 +113,7 @@ func TestWithLoader(t *testing.T) {
 		require.Nil(t, bundle)
 
 		loader := testLoader{
-			f: func(lang language.Tag, domain string) (Catalog, error) {
+			f: func(lang language.Tag, domain string) (catalog.Catalog, error) {
 				return nil, errors.New("test loader")
 			},
 		}
@@ -122,7 +123,7 @@ func TestWithLoader(t *testing.T) {
 	})
 
 	t.Run("The passed loader is set", func(t *testing.T) {
-		loader := &testLoader{f: func(lang language.Tag, domain string) (Catalog, error) {
+		loader := &testLoader{f: func(lang language.Tag, domain string) (catalog.Catalog, error) {
 			return nil, errors.New("not used")
 		}}
 		bundle, err := NewBundle(WithDomainLoader(NoDomain, loader))
@@ -167,12 +168,12 @@ func TestWithPrintFuncGenerator(t *testing.T) {
 		require.NoError(t, err)
 		require.NotNil(t, bundle)
 		require.NotNil(t, bundle.printer)
-		assert.Equal(t, 0, executionCount)
+		assert.Equal(t, 1, executionCount)
 
 		printF := bundle.printer.GetPrintFunc(language.Und)
 		result := printF("test")
 		assert.Equal(t, language.Und.String(), result)
-		assert.Equal(t, 1, executionCount)
+		assert.Equal(t, 2, executionCount)
 	})
 
 	t.Run("Nil is not a valid print function generator", func(t *testing.T) {
