+++ loader_test.go
@@ -10,6 +10,7 @@ import (
 	"github.com/stretchr/testify/require"
 	"golang.org/x/text/language"
 
+	"github.com/vorlif/spreak/catalog"
 	"github.com/vorlif/spreak/internal/util"
 )
 
@@ -236,8 +237,8 @@ func TestWithDecoder(t *testing.T) {
 		assert.NoError(t, err)
 		require.NotNil(t, fl)
 		if assert.Len(t, fl.decoder, 2) {
-			assert.IsType(t, (*poDecoder)(nil), fl.decoder[0])
-			assert.IsType(t, (*moDecoder)(nil), fl.decoder[1])
+			assert.IsType(t, catalog.NewPoDecoder(), fl.decoder[0])
+			assert.IsType(t, catalog.NewMoDecoder(), fl.decoder[1])
 		}
 
 	})
@@ -251,7 +252,7 @@ func TestWithDecoder(t *testing.T) {
 		require.NotNil(t, fl)
 		assert.Contains(t, fl.extensions, PoFile)
 		if assert.Len(t, fl.decoder, 1) {
-			assert.IsType(t, (*poDecoder)(nil), fl.decoder[0])
+			assert.IsType(t, catalog.NewPoDecoder(), fl.decoder[0])
 		}
 	})
 
@@ -264,7 +265,7 @@ func TestWithDecoder(t *testing.T) {
 		require.NotNil(t, fl)
 		assert.Contains(t, fl.extensions, MoFile)
 		if assert.Len(t, fl.decoder, 1) {
-			assert.IsType(t, (*moDecoder)(nil), fl.decoder[0])
+			assert.IsType(t, catalog.NewMoDecoder(), fl.decoder[0])
 		}
 	})
 
@@ -316,97 +317,4 @@ func TestWithResolver(t *testing.T) {
 		assert.Error(t, err)
 		require.Nil(t, fl)
 	})
-}
-
-const decodePoInvalidPluralForm = `
-msgid ""
-msgstr ""
-"Plural-Forms: invalid-pluralform\n"
-`
-
-const decodeTestData = `
-msgid "id"
-msgstr "ID"
-
-msgid "empty translation"
-msgstr ""
-
-msgid "%d day"
-msgid_plural "%d days"
-msgstr[0] "%d Tag"
-msgstr[1] "%d Tage"
-`
-
-func TestPoDecoder(t *testing.T) {
-	domain := "my-domain"
-	lang := language.German
-
-	dec := NewPoDecoder()
-	catl, err := dec.Decode(lang, domain, []byte{})
-	assert.Error(t, err)
-	assert.Nil(t, catl)
-
-	catl, err = dec.Decode(lang, domain, []byte(decodeTestData))
-	assert.NoError(t, err)
-	assert.NotNil(t, catl)
-	assert.Equal(t, lang, catl.Language())
-
-	translation, err := catl.GetTranslation("", "id")
-	assert.NoError(t, err)
-	assert.Equal(t, "ID", translation)
-
-	translation, err = catl.GetTranslation("", "xyz")
-	assert.Error(t, err)
-	assert.Equal(t, "xyz", translation)
-
-	catl, err = dec.Decode(lang, domain, []byte(decodePoInvalidPluralForm+decodeTestData))
-	assert.Error(t, err)
-	assert.Nil(t, catl)
-}
-
-func TestMoDecoder(t *testing.T) {
-	dec := NewMoDecoder()
-	decode, err := dec.Decode(language.German, "", []byte{})
-	assert.Error(t, err)
-	assert.Nil(t, decode)
-}
-
-func TestNewPoCLDRDecoder(t *testing.T) {
-	dec := NewPoCLDRDecoder()
-
-	catl, err := dec.Decode(language.German, "", []byte(decodeTestData))
-	assert.NoError(t, err)
-	require.NotNil(t, catl)
-
-	translation, errT := catl.GetPluralTranslation("", "%d day", "1.2")
-	assert.NoError(t, errT)
-	assert.Equal(t, "%d Tage", translation)
-}
-
-func TestCLDRHeader(t *testing.T) {
-	header := `
-msgid ""
-msgstr ""
-"Plural-Forms: invalid-pluralform\n"
-"X-spreak-use-CLDR: true\n"
-`
-
-	dec := NewPoDecoder()
-	catl, err := dec.Decode(language.German, "", []byte(header+decodeTestData))
-	assert.NoError(t, err)
-	require.NotNil(t, catl)
-	translation, errT := catl.GetPluralTranslation("", "%d day", "1.2")
-	assert.NoError(t, errT)
-	assert.Equal(t, "%d Tage", translation)
-}
-
-func TestGetCLDRPluralFunction(t *testing.T) {
-	pf := getCLDRPluralFunction(language.MustParse("kw"))
-	assert.Equal(t, 0, pf(0))
-	assert.Equal(t, 1, pf(1))
-	assert.Equal(t, 2, pf(22))
-	assert.Equal(t, 3, pf(143))
-	assert.Equal(t, 4, pf(161))
-	assert.Equal(t, 5, pf(5))
-	assert.Equal(t, 5, pf(1004))
 }
